Refactor imports and update documentation in notebooks
- Changed import paths for BatchSizeSuggester and AsyncBatchingMapProxy from _optimize and _proxy to _cache in batch_size.ipynb.
- Updated warning message formatting in batch_size.ipynb.
- Modified table representation in batch_size.ipynb for better readability.
- Revised FAQ entries in generate_faq.ipynb to reflect updated features and usage instructions for openaivec.
- Updated import statements in various customer support and NLP task files to import BatchResponses directly from openaivec instead of _responses.
- Adjusted example code in customer support and NLP task files to align with new import structure.
- Enhanced Spark UDF example in task initialization to use task_udf for better clarity and functionality.
