+++ src/openaivec/task/nlp/keyword_extraction.py
@@ -8,7 +8,7 @@ Example:
 
     ```python
     from openai import OpenAI
-    from openaivec._responses import BatchResponses
+    from openaivec import BatchResponses
     from openaivec.task import nlp
 
     client = OpenAI()
