+++ src/openaivec/task/nlp/translation.py
@@ -13,7 +13,7 @@ Example:
 
     ```python
     from openai import OpenAI
-    from openaivec._responses import BatchResponses
+    from openaivec import BatchResponses
     from openaivec.task import nlp
 
     client = OpenAI()
