+++ src/openaivec/task/table/fillna.py
@@ -33,7 +33,7 @@ Example:
 
     ```python
     from openai import OpenAI
-    from openaivec._responses import BatchResponses
+    from openaivec import BatchResponses
     from openaivec.task.table import fillna
 
     client = OpenAI()
